Fallback images, clean up, removing unused bits from slideshowcontent
