+++ WMFComponents/Sources/WMFComponents/Components/Year in Review/WMFSlideShow.swift
@@ -34,7 +34,6 @@ public struct WMFSlideShow: View {
                 WMFYearInReviewScrollView(
                     scrollViewContents: slideView(slide: slide),
                     hasLargeInsets: false,
-                    imageName: slides[slide].imageName,
                     gifName: slides[slide].gifName,
                     imageOverlayAccessibilityLabel: slides[slide].imageOverlayAccessibilityLabel,
                     imageOverlay: slides[slide].imageOverlay,
@@ -87,8 +86,7 @@ public struct WMFSlideShow: View {
 public protocol SlideShowProtocol {
     var title: String { get }
     var subtitle: String { get }
-    var imageName: String { get }
-    var gifName: String? { get }
+    var gifName: String { get }
     var imageOverlay: String? { get }
     var imageOverlayAccessibilityLabel: String? { get }
     var textOverlay: String? { get }
