+++ WMFComponents/Sources/WMFComponents/Components/Year in Review/WMFYearInReviewView.swift
@@ -49,7 +49,6 @@ public struct WMFYearInReviewView: View {
                     WMFYearInReviewScrollView(
                         scrollViewContents: scrollViewContent,
                         contents: { AnyView(buttons) },
-                        imageName: "intro",
                         gifName: "all-slide-01",
                         imageOverlayAccessibilityLabel: viewModel.localizedStrings.globeImageAccessibilityLabel,
                         imageOverlay: "globe_yir")
@@ -170,35 +169,5 @@ public struct WMFYearInReviewView: View {
             }
 
         }
-    }
-}
-
-struct GifImageView: UIViewRepresentable {
-    private let name: String
-    init(_ name: String) {
-        self.name = name
-    }
-    
-    func makeUIView(context: Context) -> WKWebView {
-        let webview = WKWebView()
-        webview.scrollView.contentInset = .zero
-        webview.scrollView.contentInsetAdjustmentBehavior = .never
-        webview.scrollView.bounces = false 
-        webview.isOpaque = false
-        webview.backgroundColor = .clear
-        webview.scrollView.isScrollEnabled = false
-
-        if let url = Bundle.module.url(forResource: name, withExtension: "gif"),
-           let gifData = try? Data(contentsOf: url) {
-            webview.load(gifData, mimeType: "image/gif", characterEncodingName: "UTF-8", baseURL: url.deletingLastPathComponent())
-        } else {
-            print("Error: Could not find or load gif: \(name).")
-        }
-
-        return webview
-    }
-
-    func updateUIView(_ uiView: WKWebView, context: Context) {
-        uiView.reload()
     }
 }
